[Mod] rename as vnpy_evo
