+++ CHANGELOG.md
@@ -1,3 +1,3 @@
 # 0.1.0
 
-1. The first release of VeighNa Crypto
+1. The first release of VeighNa Evo
