+++ examples/client_server/run_server.py
@@ -1,9 +1,9 @@
 from time import sleep
 
-from vnpy_crypto.event import EventEngine, Event
-from vnpy_crypto.trader.engine import MainEngine
-from vnpy_crypto.trader.ui import MainWindow, create_qapp
-from vnpy_crypto.trader.event import EVENT_LOG
+from vnpy_evo.event import EventEngine, Event
+from vnpy_evo.trader.engine import MainEngine
+from vnpy_evo.trader.ui import MainWindow, create_qapp
+from vnpy_evo.trader.event import EVENT_LOG
 
 from vnpy_binance import BinanceUsdtGateway
 from vnpy_rpcservice import RpcServiceApp
