+++ examples/simple_rpc/test_client.py
@@ -2,7 +2,7 @@ from __future__ import print_function
 from __future__ import absolute_import
 from time import sleep
 
-from vnpy_crypto.rpc import RpcClient
+from vnpy_evo.rpc import RpcClient
 
 
 class TestClient(RpcClient):
