+++ install.bat
@@ -4,5 +4,5 @@ python -m pip install --upgrade pip wheel
 ::Install prebuild wheel
 python -m pip install --extra-index-url https://pypi.vnpy.com TA_Lib==0.4.24
 
-:: Install vnpy_crypto
+:: Install vnpy_evo
 python -m pip install .
