+++ install_osx.sh
@@ -25,5 +25,5 @@ ta-lib-exists || install-ta-lib
 $python -m pip install numpy==1.23.1 --index $pypi_index
 $python -m pip install ta-lib==0.4.24 --index $pypi_index
 
-# Install vnpy_crypto
+# Install vnpy_evo
 $python -m pip install . --index $pypi_index
