+++ setup.cfg
@@ -1,6 +1,6 @@
 [metadata]
-name = vnpy_crypto
-version = attr: vnpy_crypto.__version__
+name = vnpy_evo
+version = attr: vnpy_evo.__version__
 author = VeighNa Global
 author_email = veighna@hotmail.com
 description = Core module for using VeighNa project on crypto markets.
@@ -31,7 +31,7 @@ classifiers =
     License :: OSI Approved :: MIT License
     Natural Language :: English
 project_urls =
-	Documentation = https://github.com/veighna-global/vnpy_crypto
+	Documentation = https://github.com/veighna-global/vnpy_evo
 
 [options]
 packages = find:
