+++ vnpy_evo/trader/constant.py
