+++ vnpy_evo/trader/object.py
