+++ vnpy_evo/trader/setting.py
