+++ vnpy_evo/trader/ui/ico/veighna.ico
