+++ vnpy_evo/trader/utility.py
